+++ CHANGELOG.md
@@ -1,22 +1,40 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.5.2]
+#### Fixed 
+ - Added curses.refresh in some missing code parts after a curses.clear command
+
+#### Changed 
+ - Most test code has been deleted
+
+#### Added
+ - Score coloring near level-up (TBD)
+ - More levels
+ - More shapes
+ - Beep sounds (TBD - off replit only)
+ - Proper end screen with play again
+ - Some code commenting has been added
+
+
 ## [0.4.2]
 #### Fixed 
  - Randonmess has been fixed to avoid locking into always the same shape
 
-#### Add
+#### Added
  - Proper title screen
+ - Instruction screen
+
 
 ## [0.3.1]
 #### Fixed 
  - Bug prevennting shape list and frequencies from beng updated
  - Error in not checking is we are at the last level
 
-#### Change 
+#### Changed 
  - Better randomness of shapes
  - Better control of item sequences and frequency
 
-#### Add
+#### Added
  - The cleared lines counter